convert filter options to ts
